requirements.txt
djangorestframework==3.15.2
django-contrib-postgres==0.0.1
django-environ==0.11.2
psycopg2-binary==2.9.9
django-extensions==3.2.3
ipython==8.23.0
gunicorn==22.0.0
django-cors-headers==4.4.0
safety==3.2.4
black==24.8.0
flake8==7.1.1
flake8-html==0.4.3
pre-commit==3.8.0
coverage==7.6.1
django-admin-interface==0.28.6
django-json-widget==2.0.1
qrcode[pil]==7.4.2
django-allauth==64.2.1
PyJWT==2.9.0